Update simplify-job-list.js
It was using Math.abs to determine if the date is too old for the last journal entry. Normally this is fine, but I guess the system will allow notes to be future dated somehow. So running Math.abs sets a warning if the note date is too far in the future.